requirements.txt
jinja2
markdown
pyyaml
slugify
watchdog
